Add comment remove validation

+++ WebSchool/Controllers/ApiControllers/ApiCommentsController.cs
@@ -10,16 +10,24 @@ namespace WebSchool.WebApplication.Controllers.ApiControllers
     public class ApiCommentsController : Controller
     {
         private readonly ICommentsService commentsService;
+        private readonly IUsersService usersService;
 
         public ApiCommentsController(
-            ICommentsService commentsService)
+            ICommentsService commentsService,
+            IUsersService usersService)
         {
             this.commentsService = commentsService;
+            this.usersService = usersService;
         }
 
         public async Task<IActionResult> Remove(int commentId)
         {
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if(usersService.ValidateCommentRemove(userId, commentId) == false)
+            {
+                return Json(false);
+            }
+
             var removeResult = await commentsService.RemoveAsync(userId, commentId);
             return Json(removeResult);
         }
